core: rename and organize API key classes, #TASK-4791, #TASK-4641
